backend/requirements.txt
numpy==1.26.4
faiss-cpu==1.11.0.post1
langchain==0.0.346
pymupdf==1.26.6
sentence-transformers==2.2.2
flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.5.0
requests==2.31.0
pytesseract==0.3.10
pillow==11.1.0
ollama==0.1.7
numpy==1.26.4
faiss-cpu==1.11.0.post1
langchain==0.0.346
pymupdf==1.26.6
sentence-transformers==2.2.2
flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.5.0
requests==2.31.0
pytesseract==0.3.10
pillow==11.1.0
ollama==0.1.7
